pyproject.toml
[project]
name = "transformer-ip-protection"
version = "0.1.0"
description = "Transformer IP Protection Research Project"
requires-python = ">=3.10"
dependencies = [
    "datasets>=2.0.0",
    "matplotlib>=3.5.0",
    "numba>=0.56.0",
    "numpy>=1.21.0",
    "pillow>=8.3.0",
    "pypuf>=2.2.0",
    "pyyaml>=6.0",
    "scipy>=1.7.0",
    "seaborn>=0.13.2",
    "torch>=1.9.0",
    "torchvision>=0.10.0",
    "tqdm>=4.62.0",
    "transformers>=4.20.0",
]
